+++ frontend/src/main.js
@@ -2,8 +2,5 @@ import { createApp } from 'vue'
 import App from './App.vue'
 import router from './router'
 import './assets/style.css'
-import { fetchAllAvatars } from './utils/avatarCache'
-
-fetchAllAvatars()
 
 createApp(App).use(router).mount('#app')